\name{Attribution.geometric}
\alias{Attribution.geometric}
\title{performs geometric attribution}
\usage{
  Attribution.geometric(Rp, wp, Rb, wb, Rpl, Rbl, Rbh,
    currency = FALSE)
}
\arguments{
  \item{Rp}{xts of portfolio returns}

  \item{wp}{xts of portfolio weights}

  \item{Rb}{xts of benchmark returns}

  \item{wb}{xts of benchmark weights}

  \item{Rpl}{xts, data frame or matrix of portfolio returns
  in local currency}

  \item{Rbl}{xts, data frame or matrix of benchmark returns
  in local currency}

  \item{Rbh}{xts, data frame or matrix of benchmark returns
  hedged into the base currency}
}
\value{
  This function returns the list with attribution effects
  (allocation or selection effect) including total
  multi-period attribution effects
}
\description{
  Performance attribution of geometric excess returns.
  Calculates total geometric attribution effects over
  multiple periods. Used internally by the
  \code{\link{Attribution}} function. Geometric attribution
  effects in the contrast with arithmetic do naturally link
  over time multiplicatively:
  \deqn{\frac{(1+r)}{1+b}-1=\overset{n}{\underset{t=1}{\prod}}(1+A_{t}^{G})\times\overset{n}{\underset{t=1}{\prod}}(1+S{}_{t}^{G})-1}
}
\details{
  where \deqn{A_{t}^{G}} - total allocation effect at time
  t \deqn{S_{t}^{G}} - total selection effect at time t
  \deqn{A_{t}^{G}=\frac{1+b_{S}}{1+b_{t}}-1}
  \deqn{S_{t}^{G}=\frac{1+r_{t}}{1+b_{S}}-1}
  \deqn{b_{S}=\overset{n}{\underset{i=1}{\sum}}wp_{i}\times
  rb_{i}} \deqn{b_{S}} - semi-notional fund \deqn{w_{pt}} -
  portfolio weights at time t \deqn{w_{bt}} - benchmark
  weights at time t \deqn{r_{t}} - portfolio returns at
  time t \deqn{b_{t}} - benchmark returns at time t
  \deqn{r} - total portfolio returns \deqn{b} - total
  benchmark returns \deqn{n} - number of periods

  The multi-currency geometric attribution is handled
  following the Appendix A (Bacon, 2004).

  The individual selection effects are computed using:
  \deqn{w_{pi}\times\left(\frac{1+R_{pLi}}{1+R_{bLi}}-1\right)\times\left(\frac{1+R_{bLi}}{1+b_{SL}}\right)}

  The individual allocation effects are computed using:
  \deqn{(w_{pi}-w_{bi})\times\left(\frac{1+R_{bHi}}{1+b_{L}}-1\right)}

  where \deqn{b_{SH} = \underset{i}{\sum}((w_{pi} -
  w_{bi})R_{bHi} + w_{bi}R_{bLi})} - total semi-notional
  return hedged into the base currency

  \deqn{b_{SL} = \underset{i}{\sum}w_{pi}R_{bLi}} - total
  semi-notional return in the local currency \deqn{w_{pi}}
  - portfolio weights of asset i \deqn{w_{bi}} - benchmark
  weights of asset i \deqn{R_{pLi}} - portfolio returns in
  the local currency \deqn{R_{bLi}}} - benchmark returns in
  the local currency \deqn{R_{bHi}} - benchmark returns
  hedged into the base currency \deqn{b_{L}} - total
  benchmark returns in the local currency \deqn{r_{L}} -
  total portfolio returns in the local currency The total
  excess returns are decomposed into:
  \deqn{\frac{(1+r)}{1+b}-1=\frac{1+r_{L}}{1+b_{SL}}\times\frac{1+b_{SH}}{1+b_{L}}\times\frac{1+b_{SL}}{1+b_{SH}}\times\frac{1+r}{1+r_{L}}\times\frac{1+b_{L}}{1+b}-1}

  where the first term corresponds to the selection, second
  to the allocation, third to the hedging cost transferred
  and the last two to the naive currency attribution
}
\examples{
data(attrib)
Attribution.geometric(Rp, wp, Rb, wb)
}
\author{
  Andrii Babii
}
\references{
  Christopherson, Jon A., Carino, David R., Ferson, Wayne
  E. \emph{Portfolio Performance Measurement and
  Benchmarking}. McGraw-Hill. 2009. Chapter 18-19

  Bacon, C. \emph{Practical Portfolio Performance
  Measurement and Attribution}. Wiley. 2004. Chapter 5, 8,
  Appendix A
}
\seealso{
  \code{\link{Attribution}}
}
\keyword{attribution,}
\keyword{geometric}
\keyword{linking}

