\name{Carino}
\alias{Carino}
\title{calculates total attribution effects using logarithmic smoothing}
\usage{
  Carino(rp, rb, attributions, adjusted)
}
\arguments{
  \item{rp}{xts of portfolio returns}

  \item{rb}{xts of benchmark returns}

  \item{attributions}{xts with attribution effects}

  \item{adjusted}{TRUE/FALSE, whether to show original or
  smoothed attribution effects for each period}
}
\value{
  returns a data frame with original attribution effects
  and total attribution effects over multiple periods
}
\description{
  Calculates total attribution effects over multiple
  periods using logarithmic linking method. Used internally
  by the \code{\link{Attribution}} function. Arithmetic
  attribution effects do not naturally link over time. This
  function uses logarithmic smoothing to adjust attribution
  effects so that they can be summed up over multiple
  periods. Attribution effect are multiplied by the
  adjustment factor: \deqn{A_{t}' = A_{t} \times
  \frac{k_{t}}{k}},
}
\details{
  where \deqn{k_{t} = \frac{log(1 + r_{t}) - log(1 +
  b_{t})}{r_{t} - b_{t}}}, \deqn{k = \frac{log(1 + r) -
  log(1 + b)}{r - b}}.

  In case if portfolio and benchmark returns are equal:
  \deqn{k_{t} = \frac{1}{1 + r_{t}} \deqn{A_{t}}' -
  adjusted attribution effects at period \deqn{t}
  \deqn{A_{t}} - unadjusted attribution effects at period
  \deqn{t} \deqn{r_{t}} - portfolio returns at period
  \deqn{t} \deqn{b_{t}} - benchmark returns at period
  \deqn{t} \deqn{r} - total portfolio returns \deqn{b} -
  total benchmark returns \deqn{n} - number of periods The
  total arithmetic excess returns can be explained in terms
  of the sum of adjusted attribution effects: \deqn{r - b =
  \overset{n}{\underset{t=1}{\sum}}\left(Allocation_{t}+Selection_{t}+Interaction_{t}\right)}
}
\examples{
data(attrib)
Carino(rp, rb, allocation, adjusted = FALSE)
}
\author{
  Andrii Babii
}
\references{
  Christopherson, Jon A., Carino, David R., Ferson, Wayne
  E. \emph{Portfolio Performance Measurement and
  Benchmarking}. McGraw-Hill. 2009. Chapter 19

  Bacon, C. \emph{Practical Portfolio Performance
  Measurement and Attribution}. Wiley. 2004. p. 191-193

  Carino, D. (1999) \emph{Combining attribution effects
  over time}. The Journal of Performance Measurement.
  Summer, 5–14.
}
\seealso{
  \code{\link{Attribution}} \cr \code{\link{Menchero}} \cr
  \code{\link{Grap}} \cr \code{\link{Frongello}} \cr
  \code{\link{Attribution.geometric}}
}
\keyword{arithmetic}
\keyword{attribution,}
\keyword{Carino}
\keyword{linking}
\keyword{linking,}
\keyword{logarithmic}

