\name{Grap}
\alias{Grap}
\title{calculates total attribution effects using GRAP smoothing}
\usage{
  Grap(rp, rb, attributions, adjusted)
}
\arguments{
  \item{rp}{xts of portfolio returns}

  \item{rb}{xts of benchmark returns}

  \item{attributions}{xts with attribution effects}

  \item{adjusted}{TRUE/FALSE, whether to show original or
  smoothed attribution effects for each period}
}
\value{
  returns a data frame with original attribution effects
  and total attribution effects over multiple periods
}
\description{
  Calculates total attribution effects over multiple
  periods using GRAP linking method. Used internally by the
  \code{\link{Attribution}} function. Arithmetic
  attribution effects do not naturally link over time. This
  function uses GRAP smoothing algorithm to adjust
  attribution effects so that they can be summed up over
  multiple periods Attribution effect are multiplied by the
  adjustment factor \deqn{A_{t}' = A_{t} \times G_{t}}
  where
  \deqn{G_{t}=\prod^{t-1}_{i=1}(1+R_{pi})\times\prod^{n}_{t+1}(1+R_{bi})}
  A_t' - adjusted attribution effects at period t, A_t -
  unadjusted attribution effects at period t, R_pi -
  portfolio returns at period i, R_bi - benchmark returns
  at period i, Rp - total portfolio returns, Rb - total
  benchmark returns, n - number of periods The total
  arithmetic excess returns can be explained in terms of
  the sum of adjusted attribution effects: \deqn{R_{p} -
  R_{b} =
  \sum^{n}_{t=1}\left(Allocation_{t}+Selection_{t}+Interaction_{t}\right)}
}
\examples{
data(attrib)
Grap(rp, rb, allocation, adjusted = FALSE)
}
\author{
  Andrii Babii
}
\references{
  Bacon, C. \emph{Practical Portfolio Performance
  Measurement and Attribution}. Wiley. 2004. p. 196-199 \cr
  GRAP (Groupe de Recherche en Attribution de Performance)
  (1997) \emph{Synthese des modeles d'attribution de
  performance}. Paris, Mars.\cr
}
\seealso{
  \code{\link{Attribution}} \cr \code{\link{Menchero}} \cr
  \code{\link{Carino}} \cr \code{\link{Frongello}} \cr
  \code{\link{Attribution.geometric}}
}
\keyword{arithmetic}
\keyword{attribution,}
\keyword{GRAP}
\keyword{linking}

