\name{Menchero}
\alias{Menchero}
\title{calculates total attribution effects using Menchero smoothing}
\usage{
  Menchero(rp, rb, attributions, adjusted)
}
\arguments{
  \item{rp}{xts of portfolio returns}

  \item{rb}{xts of benchmark returns}

  \item{attributions}{xts with attribution effects}

  \item{adjusted}{TRUE/FALSE, whether to show original or
  smoothed attribution effects for each period}
}
\value{
  returns a data frame with original attribution effects
  and total attribution effects over multiple periods
}
\description{
  Calculates total attribution effects over multiple
  periods using Menchero linking method. Used internally by
  the \code{\link{Attribution}} function. Arithmetic
  attribution effects do not naturally link over time. This
  function uses Menchero smoothing algorithm to adjust
  attribution effects so that they can be summed up over
  multiple periods Attribution effect are multiplied by the
  adjustment factor \deqn{A_{t}' = A_{t} \times (M
  +a_{t})}, where
  \deqn{M=\frac{\frac{1}{n}(r-b)}{(1+r)^{\frac{1}{n}}-(1+b)^{\frac{1}{n}}}}
  \deqn{a_{t} =
  \left(\frac{r-b-M\times\overset{n}{\underset{t=1}{\sum}}(r_{t}-b_{t})}{\overset{n}{\underset{t=1}{\sum}}(r_{t}-b_{t})^{2}}\right)\times(r_{t}-b_{t})}
  In case if portfolio and benchmark returns are equal
  \deqn{M = (1 + r)^\frac{n-1}{n}} \deqn{A_{t}}' - adjusted
  attribution effects at period \deqn{t} \deqn{A_{t}} -
  unadjusted attribution effects at period \deqn{t}
  \deqn{r_{t}} - portfolio returns at period \deqn{t}
  \deqn{b_{t}} - benchmark returns at period \deqn{t}
  \deqn{r} - total portfolio returns \deqn{b} - total
  benchmark returns \deqn{n} - number of periods The total
  arithmetic excess returns can be explained in terms of
  the sum of adjusted attribution effects: \deqn{r - b =
  \overset{n}{\underset{t=1}{\sum}}\left(Allocation_{t}+Selection_{t}+Interaction_{t}\right)}
}
\examples{
data(attrib)
Menchero(rp, rb, allocation, adjusted = FALSE)
}
\author{
  Andrii Babii
}
\references{
  Bacon, C. \emph{Practical Portfolio Performance
  Measurement and Attribution}. Wiley. 2004. p. 194-196

  Menchero, J. (2000) \emph{An optimized approach to
  linking attribution
}
\seealso{
  \code{\link{Attribution}} \cr \code{\link{Carino}} \cr
  \code{\link{Grap}} \cr \code{\link{Frongello}} \cr
  \code{\link{Attribution.geometric}}
}
\keyword{arithmetic}
\keyword{attribution,}
\keyword{linking}
\keyword{Menchero}

