\name{Attribution.geometric}
\alias{Attribution.geometric}
\title{performs geometric attribution}
\usage{
  Attribution.geometric(Rp, wp, Rb, wb)
}
\arguments{
  \item{Rp}{xts of portfolio returns}

  \item{wp}{xts of portfolio weights}

  \item{Rb}{xts of benchmark returns}

  \item{wb}{xts of benchmark weights}
}
\description{
  Performance attribution of geometric excess returns.
  Calculates total geometric attribution effects over
  multiple periods. Used internally by the
  \code{\link{Attribution}} function. Geometric attribution
  effects in the conrast with arithmetic do naturally link
  over time multiplicatively:
  \deqn{\frac{(1+r)}{1+b}-1=\overset{n}{\underset{t=1}{\prod}}(1+A_{t}^{G})\times\overset{n}{\underset{t=1}{\prod}}(1+S{}_{t}^{G})-1}
  , where \deqn{A_{t}^{G}} - total allocation effect at
  time t \deqn{S_{t}^{G}} - total selection effect at time
  t \deqn{A_{t}^{G}=\frac{1+b_{s}}{1+b_{t}}-1}
  \deqn{S_{t}^{G}=\frac{1+r_{t}}{1+b_{s}}-1}
  \deqn{b_{s}=\overset{n}{\underset{i=1}{\sum}}wp_{i}\times
  rb_{i}} \deqn{b_{s}} - semi-notional fund \deqn{wp_{t}} -
  portfolio weights at time t \deqn{wb_{t}} - benchmark
  weights at time t \deqn{r_{t}} - portfolio returns at
  time t \deqn{b_{t}} - benchmark returns at time t
  \deqn{r} - total portfolio returns \deqn{b} - total
  benchmark returns \deqn{n} - number of periods
}
\examples{
data(attrib)
Attribution.geometric(Rp, wp, Rb, wb)
}
\author{
  Andrii Babii
}
\references{
  Christopherson, Jon A., Carino, David R., Ferson, Wayne
  E. \emph{Portfolio Performance Measurement and
  Benchmarking}. McGraw-Hill. 2009. Chapter 18-19

  Bacon, C. \emph{Practical Portfolio Performance
  Measurement and Attribution}. Wiley. 2004. Chapter 5, 8
}
\seealso{
  \code{\link{Attribution}}
}
\keyword{attribution,}
\keyword{geometric}
\keyword{linking}

