\name{Grap}
\alias{Grap}
\title{calculates total attribution effects using GRAP smoothing}
\usage{
  Grap(rp, rb, attributions, adjusted)
}
\arguments{
  \item{rp}{xts of portfolio returns}

  \item{rb}{xts of benchmark returns}

  \item{attributions}{xts with attribution effects}

  \item{adjusted}{TRUE/FALSE, whether to show original or
  smoothed attribution effects for each period}
}
\value{
  returns a data frame with original attribution effects
  and total attribution effects over multiple periods
}
\description{
  Calculates total attribution effects over multiple
  periods using GRAP linking method. Used internally by the
  \code{\link{Attribution}} function. Arithmetic
  attribution effects do not naturally link over time. This
  function uses GRAP smoothing algorithm to adjust
  attribution effects so that they can be summed up over
  multiple periods Attribution effect are multiplied by the
  adjustment factor \deqn{A_{t}' = A_{t} \times G_{t}}{At'
  = At * Gt} where
  \deqn{G_{t}=\prod^{t-1}_{i=1}(1+R_{pi})\times\prod^{n}_{t+1}(1+R_{bi})}
  \eqn{A_{t}'}{At'} - adjusted attribution effects at
  period \eqn{t}, \eqn{A_{t}}{At} - unadjusted attribution
  effects at period \eqn{t}, \eqn{R_{pi}}{Rpi} - portfolio
  returns at period \eqn{i}, \eqn{R_{bi}}{Rbi} - benchmark
  returns at period \eqn{i}, \eqn{R_{p}}{Rp} - total
  portfolio returns, \eqn{R_{b}}{Rb} - total benchmark
  returns, \eqn{n} - number of periods The total arithmetic
  excess returns can be explained in terms of the sum of
  adjusted attribution effects: \deqn{R_{p} - R_{b} =
  \sum^{n}_{t=1}\left(Allocation_{t}+Selection_{t}+
  Interaction_{t}\right)}
}
\examples{
data(attrib)
Grap(rp = attrib.returns[, 21], rb = attrib.returns[, 22], attributions = attrib.allocation, adjusted = FALSE)
}
\author{
  Andrii Babii
}
\references{
  Bacon, C. \emph{Practical Portfolio Performance
  Measurement and Attribution}. Wiley. 2004. p. 196-199 \cr
  GRAP (Groupe de Recherche en Attribution de Performance)
  (1997) \emph{Synthese des modeles d'attribution de
  performance}. Paris, Mars.\cr
}
\seealso{
  \code{\link{Attribution}} \cr \code{\link{Menchero}} \cr
  \code{\link{Carino}} \cr \code{\link{Frongello}} \cr
  \code{\link{Attribution.geometric}}
}
\keyword{arithmetic}
\keyword{attribution,}
\keyword{GRAP}
\keyword{linking}

